+++ burn-autodiff/src/checkpoint/state.rs
@@ -1,4 +1,6 @@
-use std::any::Any;
+use std::{any::Any, collections::HashMap};
+
+use crate::graph::NodeID;
 
 /// In order to accept arbitrary node output in the same hashmap, we need to upcast them to any.
 pub(crate) type StateContent = Box<dyn Any + Send + Sync>;
@@ -23,7 +25,7 @@ impl State {
     pub(crate) fn to_state_content(&self) -> &StateContent {
         match self {
             State::Recompute { n_required: _ } => {
-                unreachable!("A child has been accessed before its parents")
+                unreachable!("Can't get state content of recompute state. A child has likely been accessed before its parents.")
             }
             State::Computed {
                 state_content,
@@ -36,7 +38,7 @@ impl State {
     pub(crate) fn into_state_content(self) -> StateContent {
         match self {
             State::Recompute { n_required: _ } => {
-                unreachable!("A child has been accessed before its parents")
+                unreachable!("Can't get state content of recompute state. A child has likely been accessed before its parents.")
             }
             State::Computed {
                 state_content,
@@ -54,5 +56,80 @@ impl State {
                 n_required,
             } => n_required.clone(),
         }
+    }
+}
+
+#[derive(new, Default)]
+/// Links [NodeID]s to their current [State]
+pub(crate) struct BackwardStates {
+    map: HashMap<NodeID, State>,
+}
+
+impl BackwardStates {
+    /// Returns the output in the [State] of the given [NodeID],
+    /// and decrements the number of times this state is required.
+    /// This function always gives ownership of the output, but will clone it if needed for further uses.
+    pub(crate) fn get_state<T>(&mut self, node_id: &NodeID) -> T
+    where
+        T: Clone + Send + Sync + 'static,
+    {
+        // Fetch the state and decrement its number of required
+        let state = self.map.remove(node_id).unwrap();
+        let remaining_n_required = state.n_required() - 1;
+
+        // Downcast the state to whatever it is supposed to be
+        // If still needed after giving ownership, we copy it back to the hashmap
+        if remaining_n_required > 0 {
+            let new_stored_state = match state {
+                State::Recompute { n_required: _ } => State::Recompute {
+                    n_required: remaining_n_required,
+                },
+                State::Computed {
+                    state_content,
+                    n_required: _,
+                } => State::Computed {
+                    state_content,
+                    n_required: remaining_n_required,
+                },
+            };
+
+            let downcasted = new_stored_state
+                .to_state_content()
+                .downcast_ref::<T>()
+                .unwrap()
+                .clone();
+
+            self.insert_state(node_id.clone(), new_stored_state);
+
+            downcasted
+        } else {
+            let downcasted = state.into_state_content().downcast::<T>().unwrap();
+            *downcasted
+        }
+    }
+
+    /// Returns a reference to the [State] of the given node
+    /// Useful when we need [State] information without needing the underlying tensor
+    pub(crate) fn get_state_ref(&self, node_id: &NodeID) -> Option<&State> {
+        self.map.get(node_id)
+    }
+
+    /// Associates a [State] to its [NodeID]
+    pub(crate) fn insert_state(&mut self, node_id: NodeID, state: State) {
+        self.map.insert(node_id, state);
+    }
+
+    pub(crate) fn save<T>(&mut self, node_id: NodeID, saved_output: T)
+    where
+        T: Clone + Send + Sync + 'static,
+    {
+        let n_required = self.get_state_ref(&node_id).unwrap().n_required();
+        self.insert_state(
+            node_id,
+            State::Computed {
+                state_content: Box::new(saved_output),
+                n_required,
+            },
+        );
     }
 }
